fix: use an absolute URI
This change adds the use of the `https` protocol to an anchor link
pointing to `tbd.website`. The addition of the protocol turns the
relative reference into an absolute reference; without this, users
clicking the link from the organization's landing page [0] are
redirected to an incorrect URI [1] which results in a GitHub error page.

[0]: https://github.com/TBD54566975
[1]: https://github.com/TBD54566975/.github/blob/main/tbd.website